Update README.md with Linux install/build details (#1)
Add in separate instructions for building Huestacean on Linux.  I wrote the instructions after successfully installing and running it on my Manjaro Linux rig.  Because I'm not the most familiar with Qt, cmake, etc.  I wasted a lot of time trying to figure out how to apply the Mac instruction about cmake_prefix to my installation.  With these changes to the README, it should be more apparent what needs to be done to install on Linux.